refactor: split naval tree builder into scoped parts (#3878)
Decompose the 675-line naval_tree_builder.dart into scope, row, and
group part files so each concern stays under the panel submodule size
target without changing naval panel tree behavior.

Refs #3878
